+++ test/scope_test.dart
@@ -99,6 +99,11 @@ void main() {
         for (var element in childB.nodes) {
           element.produce();
         }
+
+        final grandChild = childA.child('GrandChildScope')!;
+        for (final node in grandChild.nodes) {
+          node.produce();
+        }
       });
     });
 
@@ -165,21 +170,149 @@ void main() {
       test('should print scopes correctly', () {
         final root = ExampleScopeRoot(scm: Scm.testInstance);
         root.createHierarchy();
+        root.initSuppliers();
         final graph = root.graph;
         expect(
           graph,
-          'digraph unix { '
-          'subgraph cluster_ExampleRoot_1 { '
-          'label = "ExampleRoot"; subgraph cluster_ChildScopeA_2 { '
-          'label = "ChildScopeA"; ChildNodeA_3 [label="ChildNodeA"]; '
+          'digraph unix '
+          '{ subgraph cluster_ExampleRoot_1 '
+          '{ label = "ExampleRoot"; '
+          'subgraph cluster_ChildScopeA_2 '
+          '{ label = "ChildScopeA"; '
+          'subgraph cluster_GrandChildScope_4 '
+          '{ label = "GrandChildScope"; '
+          'GrandChildNodeA_5 [label="GrandChildNodeA"]; '
+          '}ChildNodeA_3 [label="ChildNodeA"]; '
           'ChildNodeB_4 [label="ChildNodeB"]; '
+          '"ChildNodeB_4" -> "ChildNodeA_3"; '
           '}subgraph cluster_ChildScopeB_3 '
           '{ label = "ChildScopeB"; '
-          'ChildNodeA_5 [label="ChildNodeA"]; '
-          'ChildNodeB_6 [label="ChildNodeB"]; '
-          '}RootA_1 [label="RootA"]; '
-          'RootB_2 [label="RootB"]; }}',
+          'subgraph cluster_GrandChildScope_5 '
+          '{ label = "GrandChildScope"; '
+          'GrandChildNodeA_8 [label="GrandChildNodeA"]; '
+          '}ChildNodeA_6 [label="ChildNodeA"]; '
+          'ChildNodeB_7 [label="ChildNodeB"]; '
+          '"ChildNodeB_7" -> "ChildNodeA_6"; '
+          '}RootA_1 [label="RootA"]; RootB_2 [label="RootB"]; '
+          '"RootA_1" -> "ChildNodeA_3"; '
+          '"RootA_1" -> "GrandChildNodeA_5"; '
+          '"RootA_1" -> "ChildNodeA_6"; '
+          '"RootA_1" -> "GrandChildNodeA_8"; '
+          '"RootB_2" -> "ChildNodeA_3"; '
+          '"RootB_2" -> "ChildNodeA_6"; '
+          '}}',
         );
+      });
+    });
+
+    group('findSupplier(name)', () {
+      test('should return the supplier with the given name or null', () {
+        final rootScope = ExampleScopeRoot(scm: Scm.testInstance);
+        rootScope.createHierarchy();
+
+        // Find a node directly contained in scope
+        final rootA = rootScope.findNode('RootA');
+        expect(rootA?.name, 'RootA');
+
+        final rootB = rootScope.findNode('RootB');
+        expect(rootB?.name, 'RootB');
+
+        // Unknown node? Return null
+        final unknownNode = rootScope.findNode('Unknown');
+        expect(unknownNode, isNull);
+
+        // Should not return child nodes
+        final childNodeA = rootScope.findNode('ChildNodeA');
+        expect(childNodeA, isNull);
+
+        // Should return nodes from parent scope
+        final childScopeA = rootScope.child('ChildScopeA')!;
+        final rootAFromChild = childScopeA.findNode('RootA');
+        expect(rootAFromChild?.name, 'RootA');
+
+        // Child nodes should find their own nodes
+        final childNodeAFromChild = childScopeA.findNode('ChildNodeA');
+        expect(childNodeAFromChild?.name, 'ChildNodeA');
+      });
+    });
+
+    group('initSuppliers()', () {
+      test(
+        'should find and add the suppliers added on createNode(....)',
+        () {
+          final scm = Scm.testInstance;
+          final rootScope = ExampleScopeRoot(scm: scm);
+          rootScope.createHierarchy();
+          rootScope.initSuppliers();
+
+          // The root node has no suppliers
+          final rootA = rootScope.findNode('RootA');
+          final rootB = rootScope.findNode('RootB');
+          expect(rootA?.suppliers, isEmpty);
+          expect(rootB?.suppliers, isEmpty);
+
+          /// The child node a should have the root nodes as suppliers
+          final childScopeA = rootScope.child('ChildScopeA')!;
+          final childNodeA = childScopeA.findNode('ChildNodeA');
+          final childNodeB = childScopeA.findNode('ChildNodeB');
+          expect(childNodeA?.suppliers, hasLength(3));
+          expect(childNodeA?.suppliers, contains(rootA));
+          expect(childNodeA?.suppliers, contains(rootB));
+          expect(childNodeA?.suppliers, contains(childNodeB));
+        },
+      );
+
+      test('should throw if a supplier is not found', () {
+        final scm = Scm.testInstance;
+        final scope = Scope.example(scm: scm);
+
+        scope.createNode<int>(
+          name: 'Node',
+          suppliers: ['Unknown'],
+          initialProduct: 0,
+          produce: (p0) {},
+        );
+
+        expect(
+          () => scope.initSuppliers(),
+          throwsA(
+            predicate<ArgumentError>(
+              (e) => e.toString().contains(
+                    'Scope "Example": Supplier with name "Unknown" not found.',
+                  ),
+            ),
+          ),
+        );
+      });
+
+      group('isAncestorOf(node)', () {
+        test('should return true if the scope is an ancestor', () {
+          final rootScope = ExampleScopeRoot(scm: Scm.testInstance);
+          rootScope.createHierarchy();
+          final childScopeA = rootScope.child('ChildScopeA')!;
+          final childScopeB = rootScope.child('ChildScopeB')!;
+          final grandChildScope = childScopeA.child('GrandChildScope')!;
+          expect(rootScope.isAncestorOf(childScopeA), isTrue);
+          expect(rootScope.isAncestorOf(childScopeB), isTrue);
+          expect(childScopeA.isAncestorOf(childScopeB), isFalse);
+          expect(childScopeB.isAncestorOf(childScopeA), isFalse);
+          expect(rootScope.isAncestorOf(grandChildScope), isTrue);
+        });
+      });
+
+      group('isDescendantOf(node)', () {
+        test('should return true if the scope is a descendant', () {
+          final rootScope = ExampleScopeRoot(scm: Scm.testInstance);
+          rootScope.createHierarchy();
+          final childScopeA = rootScope.child('ChildScopeA')!;
+          final childScopeB = rootScope.child('ChildScopeB')!;
+          final grandChildScope = childScopeA.child('GrandChildScope')!;
+          expect(childScopeA.isDescendantOf(rootScope), isTrue);
+          expect(childScopeB.isDescendantOf(rootScope), isTrue);
+          expect(childScopeA.isDescendantOf(childScopeB), isFalse);
+          expect(childScopeB.isDescendantOf(childScopeA), isFalse);
+          expect(grandChildScope.isDescendantOf(rootScope), isTrue);
+        });
       });
     });
   });